+++ tool/generate.cmd
@@ -3,7 +3,7 @@
 echo Generating C-style Win32 APIs and tests
 call dart %~dp0win32\win32api.dart
 call dart %~dp0win32\generate_ffi_jsonproto.dart 
-call dart %~dp0win32\generate_ffi_winmd.dart 
+call dart %~dp0metadata\win32.dart %~dp0..\lib\src\generated
 call dart %~dp0win32\generate_tests.dart 
 echo.
